separate out existing add_book tests
