+++ openlibrary/catalog/add_book/test_add_book.py
@@ -1,24 +1,16 @@
-from __future__ import print_function
-
 import os
 import pytest
 
 from copy import deepcopy
-from collections import defaultdict
 
 from infogami.infobase.core import Text
 
 from openlibrary.catalog import add_book
 from openlibrary.catalog.add_book import add_db_name, build_pool, editions_matched, isbns_from_record, load, RequiredField
-from openlibrary.catalog.add_book.load_book import import_author, build_query, InvalidLanguage
-from openlibrary.catalog.add_book.merge import try_merge
+from openlibrary.catalog.add_book.test_load_book import add_languages  # noqa: F401
 
-from openlibrary.catalog.merge.merge_marc import build_marc
 from openlibrary.catalog.marc.parse import read_edition
-from openlibrary.catalog.marc.marc_binary import MarcBinary, BadLength, BadMARC
-
-
-from six.moves.urllib.request import urlopen
+from openlibrary.catalog.marc.marc_binary import MarcBinary
 
 
 def open_test_data(filename):
@@ -28,20 +20,6 @@ def open_test_data(filename):
     fullpath = os.path.join(root, 'test_data', filename)
     return open(fullpath)
 
-@pytest.fixture
-def add_languages(mock_site):
-    languages = [
-        ('eng', 'English'),
-        ('spa', 'Spanish'),
-        ('fre', 'French'),
-        ('yid', 'Yiddish'),
-    ]
-    for code, name in languages:
-        mock_site.save({
-            'key': '/languages/' + code,
-            'name': name,
-            'type': {'key': '/type/language'},
-        })
 
 @pytest.fixture
 def ia_writeback(monkeypatch):
@@ -49,33 +27,6 @@ def ia_writeback(monkeypatch):
     """
     monkeypatch.setattr(add_book, 'update_ia_metadata_for_ol_edition', lambda olid: {})
 
-def test_import_author_personal_name(monkeypatch):
-    """Name order is only flipped to natural order if 'personal_name' is present.
-    """
-    monkeypatch.setattr(add_book.load_book, 'find_entity', lambda a: None)
-    result = import_author({'personal_name': 'Surname, Firstname', 'name': 'Surname, Firstname'})
-    assert result['name'] == 'Firstname Surname'
-
-def test_import_author_name_only(monkeypatch):
-    monkeypatch.setattr(add_book.load_book, 'find_entity', lambda a: None)
-    result = import_author({'name': 'Surname, Firstname'})
-    assert result['name'] == 'Surname, Firstname'
-
-def test_build_query(add_languages):
-    rec = {
-        'title': 'magic',
-        'languages': ['eng', 'fre'],
-        'authors': [{'name': 'Surname, Firstname'}],
-        'description': 'test',
-    }
-    q = build_query(rec)
-    assert q['title'] == 'magic'
-    assert q['authors'][0]['name'] == 'Surname, Firstname'
-    assert q['description'] == {'type': '/type/text', 'value': 'test'}
-    assert q['type'] == {'key': '/type/edition'}
-    assert q['languages'] == [{'key': '/languages/eng'}, {'key': '/languages/fre'}]
-
-    pytest.raises(InvalidLanguage, build_query, {'languages': ['wtf']})
 
 def test_isbns_from_record():
     rec = {'title': 'test', 'isbn_13': ['9780190906764'], 'isbn_10': ['0190906766']}
@@ -91,6 +42,7 @@ def test_editions_matched_no_results(mock_site):
     result = editions_matched(rec, 'isbn_', isbns)
     # returns no results because there are no existing editions
     assert result == []
+
 
 def test_editions_matched(mock_site, add_languages, ia_writeback):
     rec = {'title': 'test', 'isbn_13': ['9780190906764'], 'isbn_10': ['0190906766'], 'source_records': ['test:001']}
@@ -390,22 +342,6 @@ def test_build_pool(mock_site):
     pool = build_pool({'lccn': ['234'], 'oclc_numbers': ['456'], 'title': 'test', 'ocaid': 'test00test'})
     assert pool == { 'oclc_numbers': ['/books/OL1M'], 'title': ['/books/OL1M'], 'ocaid': ['/books/OL1M'] }
 
-def test_try_merge(mock_site):
-    rec = {
-        'title': 'Test item',
-        'lccn': ['123'],
-        'authors': [{'name': 'Smith, John', 'birth_date': '1980'}],
-        'source_records': ['ia:test_item'],
-    }
-    reply = load(rec)
-    ekey = reply['edition']['key']
-    e = mock_site.get(ekey)
-
-    rec['full_title'] = rec['title']
-    e1 = build_marc(rec)
-    add_db_name(e1)
-    result = try_merge(e1, ekey, e)
-    assert result is True
 
 def test_load_multiple(mock_site):
     rec = {
@@ -671,106 +607,6 @@ def test_no_extra_author(mock_site, add_languages):
     assert 'subjects' in w
     assert len(e['authors']) == 1
     assert len(w['authors']) == 1
-
-def test_don_quixote(mock_site):
-    """
-    All of these items are by 'Miguel de Cervantes Saavedra',
-    only one Author should be created. Some items have bad
-    MARC length, others are missing binary MARC altogether
-    and raise BadMARC exceptions.
-    """
-    pytest.skip("This test make live requests to archive.org")
-
-    dq = [u'lifeexploitsofin01cerv', u'cu31924096224518',
-        u'elingeniosedcrit04cerv', u'ingeniousgentlem01cervuoft',
-        u'historyofingenio01cerv', u'lifeexploitsofin02cerviala',
-        u'elingeniosohidal03cervuoft', u'nybc209000', u'elingeniosohidal11cerv',
-        u'elingeniosohidal01cervuoft', u'elingeniosoh01cerv',
-        u'donquixotedelama00cerviala', u'1896elingeniosohid02cerv',
-        u'ingeniousgentlem04cervuoft', u'cu31924027656978', u'histoiredeladmir01cerv',
-        u'donquijotedelama04cerv', u'cu31924027657075', u'donquixotedelama03cervuoft',
-        u'aventurasdedonqu00cerv', u'p1elingeniosohid03cerv',
-        u'geshikhefundonik01cervuoft', u'historyofvalorou02cerviala',
-        u'ingeniousgentlem01cerv', u'donquixotedelama01cervuoft',
-        u'ingeniousgentlem0195cerv', u'firstpartofdelig00cervuoft',
-        u'p4elingeniosohid02cerv', u'donquijote00cervuoft', u'cu31924008863924',
-        u'c2elingeniosohid02cerv', u'historyofvalorou03cerviala',
-        u'historyofingenio01cerviala', u'historyadventure00cerv',
-        u'elingeniosohidal00cerv', u'lifeexploitsofin01cervuoft',
-        u'p2elingeniosohid05cerv', u'nybc203136', u'elingeniosohidal00cervuoft',
-        u'donquixotedelama02cervuoft', u'lingnieuxcheva00cerv',
-        u'ingeniousgentlem03cerv', u'vidayhechosdeli00siscgoog',
-        u'lifeandexploits01jarvgoog', u'elingeniosohida00puiggoog',
-        u'elingeniosohida00navagoog', u'donquichottedel02florgoog',
-        u'historydonquixo00cogoog', u'vidayhechosdeli01siscgoog',
-        u'elingeniosohida28saavgoog', u'historyvalorous00brangoog',
-        u'elingeniosohida01goog', u'historyandadven00unkngoog',
-        u'historyvalorous01goog', u'ingeniousgentle11saavgoog',
-        u'elingeniosohida10saavgoog', u'adventuresdonqu00jarvgoog',
-        u'historydonquixo04saavgoog', u'lingnieuxcheval00rouxgoog',
-        u'elingeniosohida19saavgoog', u'historyingeniou00lalagoog',
-        u'elingeniosohida00ormsgoog', u'historyandadven01smolgoog',
-        u'elingeniosohida27saavgoog', u'elingeniosohida21saavgoog',
-        u'historyingeniou00mottgoog', u'historyingeniou03unkngoog',
-        u'lifeandexploits00jarvgoog', u'ingeniousgentle00conggoog',
-        u'elingeniosohida00quixgoog', u'elingeniosohida01saavgoog',
-        u'donquixotedelam02saavgoog', u'adventuresdonqu00gilbgoog',
-        u'historyingeniou02saavgoog', u'donquixotedelam03saavgoog',
-        u'elingeniosohida00ochogoog', u'historyingeniou08mottgoog',
-        u'lifeandexploits01saavgoog', u'firstpartdeligh00shelgoog',
-        u'elingeniosohida00castgoog', u'elingeniosohida01castgoog',
-        u'adventofdonquixo00cerv', u'portablecervante00cerv',
-        u'firstpartofdelig14cerv', u'donquixotemanofl00cerv',
-        u'firstpartofdelig00cerv']
-
-    bad_length = []
-    bad_marc = []
-
-    add_languages(mock_site)
-    edition_status_counts = defaultdict(int)
-    work_status_counts = defaultdict(int)
-    author_status_counts = defaultdict(int)
-
-    for ocaid in dq:
-        marc_url = 'https://archive.org/download/%s/%s_meta.mrc' % (ocaid, ocaid)
-        data = urlopen(marc_url).read()
-        try:
-            marc = MarcBinary(data)
-        except BadLength:
-            bad_length.append(ocaid)
-            continue
-        except BadMARC:
-            bad_marc.append(ocaid)
-            continue
-
-        rec = read_edition(marc)
-        rec['source_records'] = ['ia:' + ocaid]
-        reply = load(rec)
-
-        q = {
-            'type': '/type/work',
-            'authors.author': '/authors/OL1A',
-        }
-        work_keys = list(mock_site.things(q))
-        author_keys = list(mock_site.things({'type': '/type/author'}))
-        print("\nReply for %s: %s" % (ocaid, reply))
-        print("Work keys: %s" % work_keys)
-        assert author_keys == ['/authors/OL1A']
-        assert reply['success'] is True
-
-        # Increment status counters
-        edition_status_counts[reply['edition']['status']] += 1
-        work_status_counts[reply['work']['status']] += 1
-        if (reply['work']['status'] != 'matched') and (reply['edition']['status'] != 'modified'):
-            # No author key in response if work is 'matched'
-            # No author key in response if edition is 'modified'
-            author_status_counts[reply['authors'][0]['status']] += 1
-
-    print("BAD MARC LENGTH items: %s" % bad_length)
-    print("BAD MARC items: %s" % bad_marc)
-    print("Edition status counts: %s" % edition_status_counts)
-    print("Work status counts: %s" % work_status_counts)
-    print("Author status counts: %s" % author_status_counts)
 
 
 def test_same_twice(mock_site, add_languages):
